Post authorization has been implemented.

+++ backend/routes/posts.js
@@ -57,7 +57,8 @@ router.route("/")
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save().then(createdPost => {
       res.status(201).json({
@@ -80,31 +81,33 @@ router.route("/:id")
       }
     });
   })
-  .put(checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
+  .put(checkAuth, multer({ storage: storage }).single("image"), (req, res, next) => {
     let imagePath = req.body.imagePath;
     if (req.file) {
       const url = req.protocol + '://' + req.get("host");
       imagePath = url + "/images/" + req.file.filename;
     }
-    const post = new Post({
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, {
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
-    })
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, {
-      title: req.body.title,
-      content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     }).then(result => {
-      res.status(200).json({message: "Update successful!"});
+      if (result.nModified > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "Not Authorized!"});
+      }
     });
   })
   .delete(checkAuth, (req, res, next) => {
-    Post.deleteOne({_id: req.params.id})
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
       .then((result) => {
-        console.log(result);
-        res.status(200).json({message: "Post deleted!"});
+        if (result.n > 0) {
+          res.status(200).json({message: "Post deleted!"});
+        } else {
+          res.status(401).json({message: "Not Authorized!"});
+        }
       })
   });
